requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
sqlalchemy
mysqlclient
python-jose[cryptography]
passlib[bcrypt]
apscheduler
requests
stravalib
#opentelemetry-sdk
#opentelemetry-instrumentation-fastapi
#opentelemetry.exporter.otlp
python-multipart